HW1 task3 int -> float

+++ HW1/task3.py
@@ -1,8 +1,8 @@
 # Напишите программу, которая принимает на вход координаты точки (X и Y), причём X ≠ 0 и Y ≠ 0 и выдаёт номер четверти плоскости,
 # в которой находится эта точка (или на какой оси она находится).
 
-x = int(input('Enter x: '))
-y = int(input('Enter y: '))
+x = float(input('Enter x: '))
+y = float(input('Enter y: '))
 if x == 0 or y == 0:
     print('Enter values not equal 0')
 else:
